requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
google-generativeai>=0.3.0
langsmith>=0.0.69
langgraph>=0.0.40
python-dotenv>=1.0.0
chromadb>=0.4.0
nest-asyncio>=1.5.0
fastmcp>=0.4.1
